Fixed some "!SMS$IGNORE" problems resulting from the fact the the ignore directive must start in column 1.
The writing of FLIP_ERR was more complicated than I thought. 
The file FLIP_ERR was being opened in a serial region so only the root was opening it so various diagnostic writes to FLIP_ERR from other processors were going into the bit bucket (lost). 
I changed the open to open for all processors so now I get a FLIP_ERR that has the UTIME written by the root and various diagnostic messages written by other processors. 
An 80p run of the 170x80 case gave 479 "Convergence failure" diagnostics in FLIP_ERR.  

It's b4b with the original run and takes the same time as the run without FLIP_ERR.